AddImage created, update profile photo
